Housekeeping: Moved celebritybucks api request to controllers folder
