+++ seeds/fameData.js
@@ -1,11 +1,11 @@
 const { Fame } = require('../models');
-var { getPeople } = require('../api');
-// const { placeholder } = require('sequelize/types/lib/operators');
+var { getPeople } = require('../controllers/api');
 
-// function to map people array to expected values of fame array
-// Get the people array from the axios api request
+// Map people array to expected values of fame array
 async function createPeople() {
+    // Get the array of people from the CelebrityBucks api call
     const people = await getPeople();
+    // Map the contents of the people array objects to the expected values of Fame objects
     const celebrities = people.map(celebrity => {
         const newCeleb = {
             id: parseInt(celebrity.celebId),
@@ -17,13 +17,15 @@ async function createPeople() {
     console.log("celebrities below");
     console.log(celebrities);
     
+    // Cycle through the celebrities array and create a Fame record for each object
     var fameData = celebrities.map(async (celebrity) => {
         return await Fame.create(celebrity)
     });
-    
+    // Return the fameData array only after information has been added to it (promise complete)
     fameData = await Promise.all(fameData);
     console.log("Fame data creation", fameData);
     return fameData;
 }
 
+// Export for use in seed.js
 module.exports = { createPeople };
